ci: Kill tests after 60s using nextest (#2900)
## Description

By using nextest to kill slow tests we get logs when a test is killed.
This is better than the github timeout kicking in as then we get no
feedback on what was stuck.

Tests are now marked as slow after 20s and killed after 60s.


## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

discovery::tests::endpoint_discovery_combined_wrong_only seems like it
is a slow tests needing about 30s to complete. To be fair I think it's
fine to call that out as a slow test, so I'm fine with that.

It is possible that this 60s timeout will introduce some new flakyness
to the slow CI machines. We'll have to see and potentially tweak this a
bit.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.