chore: adapt to latest main in iroh-gossip and iroh-docs (#2936)
## Description

adapt to latest main in iroh-gossip and iroh-docs

still not sure if handle_rpc_request should take an `Arc<Self>`, but for
now just make it work

## Breaking Changes

none

## Notes & open questions

Should all fns that take self take an `Arc<Self>` in a protocol handler?
The Arc kinda infects the entire code base. But maybe we have to do it.

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.