fix(ci): try to reuse msys2 dep (#2956)
## Description

According to
https://github.com/msys2/setup-msys2?tab=readme-ov-file#release this
should try to reuse an existing msys2 installation (though seems to
reference a default fallback to the one installed on the GH hosted
runners, unsure if it works for self hosted)

Edit: seems to be working

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.