feat(iroh-base, iroh-net-report)!: intro net-report as a crate (#2921)
## Description

Introduces net-report as its own crate. For this, introduces a new
feature flag `relay` in `iroh-base` for the relay topology types.

- in `iroh-base`, the `RelayUrl` is moved to its own file. It can still
be used as before, this is not a breaking change.
- `RelayMap`, `RelayNode` are moved to `iroh-base` under the `relay`
feature flag.
- `RelayMode` is moved to `endpoint`. See the NOTES section about this.
- net-report now has metrics as a feature flag. I'm not sure how is that
this was not feature flagged before but still worked.
- ping is moved without changes from `iroh-net` to `iroh-net-report`.
This is the only place where it's used so it should have probably been
part of `netcheck` as a module from the beginig.
- Adds missing license files


## Breaking Changes

- `iroh-net`'s `NetcheckMetrics` are now called `NetReportMetrics`

## Notes & open questions

- `RelayMode` remains in `iroh-net` instead of moving to `iroh-base`
with other relay related because this is coupled with the staging and
prod configuration of relay urls. Since this is more configuration than
concept I think it's best to keep it in iroh-net.
- The crate is called `iroh-net-report` because -at least currently- it
depends on relay concepts. For example, the tests depend heavily on the
internal order of the `RelayMap`. Can we change this? yes. Should we
change this? we can discuss it. Is this PR the place for any of that?
No.
- The api needs some love. Some types that are supposed to be internal
are used by dependents (in this case `iroh-net`) It's not terrible but
it could be improved.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] ~Tests if relevant.~
- [ ] ~All breaking changes documented.~