ref(iroh-net)!: Make Endpoint::close not consume self (#2882)
## Description

The endpoint can be freely cloned, so consuming `self` on shutdown is
somewhat fruitless. This changes it to take `&self` like a normal
method.

## Breaking Changes

- `Endpoint::close` now takes `&self` rather than `self`. This can in
some situations mean an existing (clone of an) endpoint might be dropped
too early as a temporary variable.

## Notes & open questions

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.

---------

Co-authored-by: iacore <noreply+gpg-stub@1a-insec.net>
Co-authored-by: Floris Bruynooghe <flub@devork.be>
Co-authored-by: Floris Bruynooghe <flub@n0.computer>