refactor(iroh): Bypass magicsock::Actor for datagrams from the relay
Datagrams received via the relay used to be sent to the
`magicsock::Actor` for processing (handling DISCO packets and adding
RecvMeta).  After processing they were put on a channel from which
`AsyncUdpSocket::poll_recv` would read them.

The major problem was that the relay actor did not handle the
backpressure well: the channel it was placing the processed datagrams
on is smaller than its inbox and it did not want to drop
packets (which is good, dropping packets at this point is not
helpful).  So the entire actor blocked, which has negative effects all
around.

This change connects a channel directly between the `ConnectedRelay`
actor which receives the packets and the `AsyncUdpSocket::poll_recv`.
The processing which was happening in the `magicsock::Actor` now
happens inside `poll_recv`.  This is fine because:

- The processing needs access to the `MagicSock` inners, like the
  `NodeMap`.  This required accessing the same mutexes that
  `AsyncUdpSocket::poll_recv` was already accessing anyway.  So this
  path is now slowed down.

- The alternative was moving the processing to the `ConnectedRelay`
  actor, which would have added potential more conention on the
  `MagicSock` mutexted-inners.

To support this there are a few additional  changes:

- The channel between the `AsyncUdpSocket::poll_recv(_relay)` and the
  `ConnectedRelay` actor has been abstracted into the
  `RelayRecvChannel`.  This takes care of the wakers, which now do not
  need to be separate fields on the `MagicSock`, reducing the
  `MagicSock` state and risk of using the wakers wrongly.

- The `ActiveRelay` actor is renamed to `ConnectedRelay`.  This
  improves readability.

- The `PacketSplitIter` is moved into the `relay_actor` module, this
  is where the corresponding `PacketizeIter` lives and makes the two
  much better logically linked.

- ReceviedMessage::ReceivedPacket::source is renamed into
  remote_node_id.