chore(ci): reduce sccache size (#2988)
## Description

Im not convinced this is actually helping much on our self hosted
runners.
Also we've started bumping into disk size issues on some machines
because of our recent proliferation of repos.

I didn't remove it entirely so we can observe it for a bit, but would
potentially like to follow up and drop it unless we want to rework this
with a large central cache server for our CI boxes.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.