fix(iroh-dns-server): remove accidental blocking from store (#2985)
## Description

I found this during investigations of #2972. It turned out that both
blocking locks and blocking IO calls are being made in the dns servers
store implementation

## Breaking Changes

None

## Notes & open questions

This might be not optimal, but it is the safest way to get rid of
blocking the whole runtime for now.

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.