refactor(iroh): Remove unused rate limiter
This was a rate limiter which seems like the intention was to slow
down how fast it sends to a particular relay server.  This is the
wrong place to rate-limit.  If the server limits that it should read
slower from the TCP stream (it does this now!) and that blocks up the
TCP-stream and in turn slows down the relay client.  In no world does
a socket internally rate-limit to one of it's destinations, that's not
how sockets work.  And this client exists for the MagicSock.