refactor(iroh): Remove unused rate limiter (#3007)
## Description

This was a rate limiter which seems like the intention was to slow
down how fast it sends to a particular relay server.  This is the
wrong place to rate-limit.  If the server limits that it should read
slower from the TCP stream (it does this now!) and that blocks up the
TCP-stream and in turn slows down the relay client.  In no world does
a socket internally rate-limit to one of it's destinations, that's not
how sockets work.  And this client exists for the MagicSock.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.