chore: (Deps) Update swarm-discovery to avoid idna <= 0.5 dep (#3025)
## Description

<!-- A summary of what this pull request achieves and a rough list of
changes. -->
- Updates to latest `swarm-discovery` alpha, so that `cargo deny` is
happy again

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->
None, swarm-discovery isn't exposed in the API.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->
It might be questionable if we want to depend on *more* alphas, but
IMHO, this case is fine to avoid a security vulnerability.
Hopefully this is the last alpha. The next version of hickory we depend
on should be 0.25.

## Change checklist

- [x] Self-review.
- ~~[ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.~~
- ~~[ ] Tests if relevant.~~
- [x] All breaking changes documented.