refactor(iroh): rename ConnectedRelayActor to ActiveRelayActor
Yes, that's (almost) back to what it was called...

The name has grown on me and it is better because the relay is not
always connected.  This does also fit better with what it should be
doing: keep the connection alive if possible.  Currently this
functionality is mixed up in the RelayActor itself.