docs(*): Use doc-auto-cfg feature
This switches from the doc-cfg feature which needs an explicit
annotation to the doc-auto-cfg feature which does not need this.  We
already had several places where the docs and actual features were out
of sync.  So hopefully this will work better.

doc-auto-cfg does not seem perfect and may miss some things.  E.g. one
difference I noticed is that it does not seem to surface the "test"
feature when we do `any(test, feature = "something")`.  But I'm
thinking that might be a reasonable price to pay here.