refactor(iroh): Newtype the packet sent over relay servers
There is a difference between the datagrams that the MagicSock sends
and the packets that contain datagrams that gets sent to relay
servers.  But this was all just Bytes.

This newtypes the packets explicitly and makes it a sibling of the
existing RelayRecvDatagram.  Now the PacketizeIter at least does not
both consume and produce the same type, because they're not the same
things.