refactor(iroh): rename ConnectedRelayActor to ActiveRelayActor (#3027)
## Description

Yes, that's (almost) back to what it was called...

The name has grown on me and it is better because the relay is not
always connected. This does also fit better with what it should be
doing: keep the connection alive if possible. Currently this
functionality is mixed up in the RelayActor itself.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

- Just a plain rename.
- Base is `flub/relay-send-channel-1`, be aware of github when merging
things.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.