refactor(iroh): Newtype the packet sent over relay servers (#3030)
## Description

There is a difference between the datagrams that the MagicSock sends and
the packets that contain datagrams that gets sent to relay servers. But
this was all just Bytes.

This newtypes the packets explicitly and makes it a sibling of the
existing RelayRecvDatagram. Now the PacketizeIter at least does not both
consume and produce the same type, because they're not the same things.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

This clones the RelayUrl (usually) twice per packet sent now, instead of
(usually) once. That's the only thing I don't like about this really.

I still wonder whether RelayUrl should not be some sort of interned and
cloned thing.

Watch out, the base of this PR is `flub/active-relay-again`.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.