refactor(iroh, iroh-relay): JoinSet disabling in tokio::select!
This changes the disabling of JoinSet::join_next branches in
tokio::select! calls to rely on pattern matching alone.  This changes
nothing to the logic executing in the select! call, but often results
in less code.

A branch is disabled if the pattern does not match, and the remaining
futures are still awaited.  Once the branch is disabled the logic of
tokio::select! is exactly the same regardless of whether it was
disabled because of an if condition or because of a mismatched pattern.