refactor(iroh, iroh-relay): JoinSet disabling in tokio::select! (#3052)
## Description

This changes the disabling of JoinSet::join_next branches in
tokio::select! calls to rely on pattern matching alone.  This changes
nothing to the logic executing in the select! call, but often results
in less code.

A branch is disabled if the pattern does not match, and the remaining
futures are still awaited.  Once the branch is disabled the logic of
tokio::select! is exactly the same regardless of whether it was
disabled because of an if condition or because of a mismatched pattern.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.