feat(net-report)!: add `net_report::Options` to specify which probes you want to run (#3032)
## Description

Breaking change that allows you to specify which probe protocols you
would like to run.

Really written so it is easier to test QUIC address discovery on `iroh
doctor`, but this also allows us to test specific protocols in
isolation.

## Breaking Changes

- `iroh-net-report`
    - added
        - `net_report::Client::get_report_with_options`
    - changed
- `net_report::Client::get_report_channel` now takes an `opts:
net_report::Options`
- `net_report::Client` will no longer bind `UdpSocket`s when one is not
provided for both STUN over IPv4 or STUN over IPv6.

## Notes & open questions

I'm open to adjustments on the `Options`. The whole "we build a UDP
socket for you if one isn't provided" bit means that I couldn't rely on,
for example, `stun_sock_v4 == None` to know whether the user wanted to
run a stun ipv4 probe, so I needed the `protocols:
BTreeSet<ProbeProtocol>` field. I think this is a decent compromise
between annoying and useful, but would love feedback.

Edit:
Adjusted this so that the Options DO correlate to the protocols we want
to run. This means that we will no longer bind a UdpSocket for you if
one is not provided.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.

---------

Co-authored-by: “ramfox” <“kasey@n0.computer”>