ci(*): Run flaky tests with --verbose and use ci profile
Also fix building of tests to use --profile ci which is used to run
the tests.  The test run did not trigger a re-build.  So mosly likely
the profile is not encoded in the build artifacts and we were running
tests on the normal dev profile rather than the ci profile.