ci(*): Run flaky tests with --verbose and use ci profile (#3063)
## Description

Also fix building of tests to use --profile ci which is used to run
the tests.  The test run did not trigger a re-build.  So mosly likely
the profile is not encoded in the build artifacts and we were running
tests on the normal dev profile rather than the ci profile.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.