fix(iroh-relay): removes deadlock in `Clients` (#3099)
## Description

Remove the deadlock that was a bit hidden due to `DashMap`.

Added `client` parameter to `unregister` to explicitly drop before
attempting to call into the `clients` `DashMap` again.

## Notes & open questions

The `warn` log for stream termination seems a little fear-mongering, but
I'm not sure the best way to "downgrade" this, as we seem to rely on
this error in tests like `endpoint_relay_connect_loop`. Ended up leaving
it as is.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.

---------

Co-authored-by: “ramfox” <“kasey@n0.computer”>
Co-authored-by: Friedel Ziegelmayer <me@dignifiedquire.com>