docs(iroh): Update discovery docs, mostly StaticProvider
Updates a bunch of docs around StaticProvider and links from the right
places.

This also removes the ability to build docs without using
--all-features.  The conditionals made things build but are not
correct, items that should be documented on docs.rs were not.  This
goes back to the simpler solution and documents for everyone how they
should build docs.