tests(iroh): Add some context to test errors (#3066)
## Description

The flaky tests are being super weird and making no sense.  This is
clutching at straws for any hint.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

/cc @ramfox 

Two slightly odd things I spotted while looking at this:

- The use of 240.0.0.1 as "bad" IP address returned in discovery. Maybe
one from the test nets in RFC 5737 would be better?
- The endpoints accept a connection and immediately drop it. This could
potentially send the CONNECTION_CLOSE before the ACK from the connection
resulting in the connect call returning an error (BUT THAT WOULD BE A
NORMAL READABLE TEST FAILURE). This probably doesn't happen because it's
all local and very fast and Quinn has already ACKed the connection
before the code manages to drop it. Still, essentially racy code.

Fine, I added a fix for the 2nd.  But it's not the issue.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.