feat: Make `Endpoint::close` infallible (#3112)
## Description

<!-- A summary of what this pull request achieves and a rough list of
changes. -->
The only source for a `Result::Err` in `Endpoint::close` was a failed
magic sock actor `Shutdown` message send.

This send should basically shutdown the actor, and sending only fails
when the actor dropped the receiver, which only happens when the actor
already shut down, gracefully or otherwise. In any case, it can't react
to that anyways and it's not worth surfacing that error to the user.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->
- `iroh::Endpoint::close`'s future is now infallible, instead of
returning a `Result`.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- ~~[ ] Tests if relevant.~~
- [x] All breaking changes documented.